+++ frontend/src/Dumps.vue
@@ -1,0 +1,3 @@
+<template>
+    <back></back>
+</template>
